requirements.txt
streamlit==1.24.0
qrcode==7.4.2
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3